Fix #1407: Wire PerfTrace into transaction commit path
Moves PerfTrace/perf_time! from engine to core crate (resolves circular
dependency) and wraps validation, WAL append, and storage apply phases
in commit_with_wal_arc() with zero-cost timing. Feature propagation:
root → engine → concurrency → core via perf-trace feature flag.

Invariants verified: ACID-001, ACID-002, ACID-003, ACID-004, MVCC-003, ARCH-002
