+++ crates/concurrency/Cargo.toml
@@ -8,6 +8,10 @@ license.workspace = true
 repository.workspace = true
 publish = false
 
+[features]
+default = []
+perf-trace = ["strata-core/perf-trace"]  # Enable per-operation timing in commit path
+
 [dependencies]
 strata-core = { path = "../core" }
 strata-storage = { path = "../storage" }
