+++ crates/core/src/lib.rs
@@ -18,6 +18,7 @@
 pub mod branch_types; // Branch lifecycle types
 pub mod contract; // contract types
 pub mod error;
+pub mod instrumentation; // Performance tracing (feature-gated)
 pub mod limits; // Size limits for keys, values, and vectors
 pub mod primitives; // primitive types (Event, State, Vector, JSON types)
 pub mod traits;
